Extract bed repository helpers into dedicated module
